Extract `snac-renderer-V2` library from refactor_proto.
