+++ src/libs/snac-renderer-V2/snac-renderer-V2/Json.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <nlohmann/json.hpp>
+
+using Json = nlohmann::json;
