+++ src/libs/snac-renderer-V2/snac-renderer-V2/Logging-init.h
@@ -1,0 +1,22 @@
+#pragma once
+
+
+namespace ad::renderer {
+
+
+class LoggerInitialization
+{
+private:
+    // __declspec is not forcing the export for static libs, apparently.
+    /*__declspec(dllexport)*/ static const LoggerInitialization gInitialized;
+    LoggerInitialization();
+};
+
+
+} // namespace ad::renderer
+
+
+extern "C"
+{
+void ad_renderer_loggerinitialization();
+}
