+++ src/libs/snac-renderer-V2/snac-renderer-V2/files/BinaryArchive.h
